Refactor DeltaQuery classes: reorganize packages, add `ListPartitions` and `Reconnect` classes, and improve structure for queries and subscriptions.
