+++ client/src/main/java/io/lionweb/client/delta/messages/queries/ListPartitions.java
@@ -1,0 +1,11 @@
+package io.lionweb.client.delta.messages.queries;
+
+import io.lionweb.client.delta.messages.DeltaQuery;
+import org.jetbrains.annotations.NotNull;
+
+public class ListPartitions extends DeltaQuery {
+
+  public ListPartitions(@NotNull String queryId) {
+    super(queryId);
+  }
+}
